pyproject.toml
[project]
name = "salesagent-openaisdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "openai>=2.32.0",
    "openai-agents>=0.14.7",
    "python-dotenv>=1.2.2",
    "sendgrid>=6.12.5",
]
